refactor: move `Calendar` to atom
